eventscripts: Move NFS reconfigure to "ipreallocated" event
Doing this in the "monitor" event is unsafe because it causes the node
health status to flip-flop.  At the moment when a node goes unhealthy
it is failed out, IPs are released and the monitor event handles the
reconfigure, returning 0 even though the service failure is
unresolved.

This change was made in the master branch a long time ago.

Signed-off-by: Martin Schwenke <martin@meltin.net>