Fix the build after backporting f3bf2ab61f8dbbc806ec23a68a87aaedd458e712.
This patch (keeping track of public IP assignment in recovery daemon)
which was backported to 1.0.0114 as 9640e2bb889bd99389d9fb247191a19785a75104
renamed "struct _trbt_tree_t" to "struct trbt_tree".

In master, this patch came before the introduction of the delete queue
to the db context. So in the 1.0.114 branch we need to fix up afterwards.

Signed-off-by: Michael Adam <obnox@samba.org>